Renaming status package to reporter
