+++ assembly/compliance-tool/src/test/java/org/eclipse/sw360/antenna/frontend/compliancetool/main/AntennaComplianceToolTest.java
@@ -12,7 +12,7 @@ package org.eclipse.sw360.antenna.frontend.compliancetool.main;
 
 import org.eclipse.sw360.antenna.frontend.compliancetool.sw360.SW360Configuration;
 import org.eclipse.sw360.antenna.frontend.compliancetool.sw360.exporter.SW360Exporter;
-import org.eclipse.sw360.antenna.frontend.compliancetool.sw360.status.*;
+import org.eclipse.sw360.antenna.frontend.compliancetool.sw360.reporter.*;
 import org.eclipse.sw360.antenna.frontend.compliancetool.sw360.updater.SW360Updater;
 import org.eclipse.sw360.antenna.sw360.workflow.generators.SW360UpdaterImpl;
 import org.junit.Rule;
